`sparse.csgraph.dijkstra` implements the dijkstra algorithm to find the shortest path of a graph.
The comment in the code insists that the total time complexity of running it for all source vertices is $O(N(Nk + N\log(N)))$ where $N$ is the number of nodes and $k$ is the average number of connected edges per node.
However, it turned out that the worst case complexity is being $\Theta(N^3)$ in the current implementation. Star graph(graph with edges connecting node #0 and node #i for all 1 <= i < n) is one of the worst cases.
This is due to an incomplete implementation of Fibonacci heap.
<details>
<summary>Details of the implementation flaw</summary>
The number of children of a node in Fibonacci heap can be temporarily very large after repeated calls of `insert_node` because `insert_node` does nothing other than just appending a new child.<br>
`remove_min` iterates over these large number of children but it does not impact the overall complexity because `remove_min` rearranges those nodes so that there is no more such problematic situation around the nodes. In order for `remove_min` to iterate over a large number of children again, a large number of `insert_node` is needed, so amortized analysis works.  
However, this also means that we should not touch children when we do not resolve the high-rank vertex.<br>
In the current implementation, leftmost_sibling and rightmost_sibling(which both iterate over the children) are called in other context, causing the time complexity to bump.<br>
</details>
<details>
<summary>Details of the solution</summary>
Currently, node.children points to an arbitrary child of the node. In the new implementation, it always points to the leftmost child of the children. In other words, node.children.left_sibling is always NULL.<br>
Also, heap.min_node is also kept at the leftmost end of the linked list of roots.<br>
This way, there is no need to iterate over the children of a node in places other than remove_min.
</details>
The debugging function at the bottom of the file is also changed so that it is compatible with the new implementation(and with Python 3 !).<br>
Star graphs are added to the benchmark file, and a typo in the file name is fixed.

Here is the benchmark result of before/after on i5-10600k, VirtualBox Ubuntu 20.04. The new implementation is substantially faster with format = 'star' and slightly faster with format = 'random' which is the existing benchmark.
```
Before:
[100.00%] ··· ===== =============== ============= ================ ==============
              --                          min_only / format
              ----- -------------------------------------------------------------
                n    True / random   True / star   False / random   False / star
              ===== =============== ============= ================ ==============
                30      178±2μs        143±1μs       193±0.9μs       152±0.5μs
               300      417±5μs        204±1μs      5.05±0.02ms     1.80±0.01ms
               900    2.16±0.01ms      700±3μs       112±0.3ms       50.2±0.2ms
              ===== =============== ============= ================ ==============
After:
[100.00%] ··· ===== =============== ============= ================ ==============
              --                          min_only / format
              ----- -------------------------------------------------------------
                n    True / random   True / star   False / random   False / star
              ===== =============== ============= ================ ==============
                30      176±2μs       144±0.3μs      191±0.9μs       147±0.6μs
               300      380±2μs        163±1μs      4.08±0.03ms       522±3μs
               900    1.99±0.03ms      201±1μs       95.9±0.5ms     3.50±0.03ms
              ===== =============== ============= ================ ==============
```