@rgommers 2022-07-26
> I'm not sure if this belongs to MAINT or BUG.

I'd say ENH for significant performance improvements. This looks quite good, thanks @windows-server-2003