@jjerphan 2022-07-27
@rgommers: the "curious me" definitely wants to jump in it, yet the "busy me" is "Wait, look at you todo-max-heap, Julien". :smile: 

I might start on my free-time, probably in a few weeks?

I have a few PRs I also have started reviewing for SciPy, and it's probably better than I work toward having them merge first, IMO.