@windows-server-2003 2022-08-27
@jjerphan I know you are busy, but look at the 14x speedup in the most fundamental function in graph area!  It will surely improve scipy a lot!