@jjerphan 2022-08-27
Hi @windows-server-2003,

I appreciate the speed-up brought by your PR. I have it on my list of reviews, unfortunately I am currently busy with other things at the moment. Thank you for your patience.

(Also, I like your GitHub handle.)