@windows-server-2003 2022-08-30
Okay, always thanks for working on scipy and related things!