@windows-server-2003 2022-09-12
I just remembered that C++ STL can be used in Cython
std::priority_queue should replace the Fibonacci Heap, and

 - It reduces the maintenance cost
 - priority_queue doesn't have decrease_val but we can just push another node instead
 - I suspect priority_queue is faster(needs benchmark though)

If it's okay to use C++ STL in scipy, I want to test it