@windows-server-2003 2022-09-12
Here's the results of the benchmark(3 runs each, I'm only able to use the one-time benchmark `dev.py bench` because I haven't found the proper way of using libcpp from Cython yet)

<details>
<summary>Benchmark results</summary>

```
FibonacciHeap
[ 50.00%] ··· ===== =============== ============= ================ ==============
              --                          min_only / format
              ----- -------------------------------------------------------------
                n    True / random   True / star   False / random   False / star
              ===== =============== ============= ================ ==============
                30      387±0μs        496±0μs        390±0μs         513±0μs
               300      618±0μs        517±0μs        4.19±0ms        946±0μs
               900      2.48±0ms       553±0μs        90.2±0ms        4.38±0ms
              ===== =============== ============= ================ ==============
              ===== =============== ============= ================ ==============
                30      382±0μs        2.08±0ms       1.26±0ms        843±0μs
               300      1.04±0ms       896±0μs        7.31±0ms        946±0μs
               900      2.53±0ms       560±0μs        90.6±0ms        4.42±0ms
              ===== =============== ============= ================ ==============
              ===== =============== ============= ================ ==============
                30      411±0μs        492±0μs        400±0μs         505±0μs
               300      658±0μs        1.15±0ms       8.66±0ms        933±0μs
               900      2.57±0ms       560±0μs        90.0±0ms        4.42±0ms
              ===== =============== ============= ================ ==============

std::priority_queue
[ 50.00%] ··· ===== =============== ============= ================ ==============
              --                          min_only / format
              ----- -------------------------------------------------------------
                n    True / random   True / star   False / random   False / star
              ===== =============== ============= ================ ==============
                30      413±0μs        1.56±0ms       786±0μs         1.07±0ms
               300      1.18±0ms       657±0μs        3.06±0ms        1.21±0ms
               900      2.17±0ms       574±0μs        41.9±0ms        5.21±0ms
              ===== =============== ============= ================ ==============
              ===== =============== ============= ================ ==============
                30      400±0μs        2.18±0ms       1.76±0ms        854±0μs
               300      1.04±0ms       890±0μs        8.38±0ms        1.58±0ms
               900      2.21±0ms       600±0μs        42.0±0ms        5.15±0ms
              ===== =============== ============= ================ ==============
              ===== =============== ============= ================ ==============
                30      453±0μs        503±0μs        428±0μs         1.36±0ms
               300      1.54±0ms       570±0μs        2.51±0ms        2.82±0ms
               900      2.30±0ms       596±0μs        42.0±0ms        5.20±0ms
              ===== =============== ============= ================ ==============
```

</details>

Neither of the two constantly outperforms the other, but I think std::priority_queue is a better choice because

 - It shows significant(50%+) improvement in one of the tests while Fibonacci Heap is only 20% faster in the best case(among these testcases)
 - It would reduce 300 lines of code and gives better maintenability

What do you think?