@jjerphan 2022-09-12
Thanks for performing the comparison! 
By default `std::priority_queue` is backed by a `std::vector` seen as a binary heap. This option simpler and more efficient in practice than Fibonacci Heap when the number of elements is small to medium due to data locality mainly.

IIRC, the theoretical better complexity of Fibonacci Heaps really arises for use-case with large number of samples.

Hence, I would rerun the benchmark for `n in [10_000, 100_000, 1_000_000]` (if possible).

If `std::priority_queue` is competitive, I think it makes sense to replace Fibonacci Heap by `std::priority_queue` for eased maintenance. But it's up to SciPy's maintainers to decide.

---

As for using `std::priority_heap` from Cython, you should be able to `cimport` symbols via [the declaration set by Cython](https://github.com/cython/cython/blob/efdf996dced075682d86b7e53a12b56b5ebfea38/Cython/Includes/libcpp/queue.pxd):

```cython
from libcpp.priority_queue cimport queue, priority_queue
```

and use it similarly to as you would in C++, see [this test in the Cython code-base](https://github.com/cython/cython/blob/c25c87d71107e634162302f7f61a119eff539a48/tests/run/cpp_stl_cpp11.pyx#L61-L73).

You will also need to add support for C++ when creating the extension, that is you need to modify:

https://github.com/scipy/scipy/blob/06da9ee65cdee4a8aa036767489bfc5a030527f4/scipy/sparse/csgraph/setup.py#L10-L12

into:

```python
    config.add_extension('_shortest_path',
         sources=['_shortest_path.c'],
         include_dirs=[numpy.get_include()],
         language="c++",
)