@jjerphan 2022-09-12
Thank you for the new results, @windows-server-2003.

I guess another PR can be created to replace Fibonacci Heaps by `std::priority_queue`s, maintaining this one open in the meantime.