@tylerjereddy 2022-11-24
As far as I can tell, this PR should provide a performance improvement on its own. I read through all the comments above, and I don't really see any blockers, just suggestions for even more improvements.

Let me see if I can apply some of the suggestions, rebase (this hasn't flushed through CI since July), and push + run benchmarks locally.